requirements.txt
Django==4.2
djangorestframework==3.14
drf-spectacular>=0.15.1,<0.16
requests==2.31
beautifulsoup4==4.12